data: Add keyring to /usr/share/keyrings
The preferred location for keyrings is /usr/share/keyrings. The trixie
manpage for sources.list says:-

•   Signed-By (signed-by) is an option to require a repository to pass
           apt-secure(8) verification with a certain set of keys rather
	   than all trusted keys apt has configured. It is specified as
	   a list of absolute paths to keyring files (have to be
	   accessible and readable for the _apt system user, so ensure
	   everyone has read-permissions on the file) and fingerprints
	   of keys to select from these keyrings.
           The recommended locations for keyrings are /usr/share/keyrings
	   for keyrings managed by packages, and /etc/apt/keyrings for
           keyrings managed by the system operator. If no keyring files
	   are specified the default is the trusted.gpg keyring and all
	   keyrings in the trusted.gpg.d/ directory.

The keyring will still be available in /etc/apt/trusted.gpg.d as OBS is
still using this location in directions for setting up a repository for
use.

For the new key the pgp extension is used as preferred by Debian and
because it is pgp, not just gpg. Link [1].

Link: [1] https://wiki.debian.org/DebianRepository/UseThirdParty
Signed-off-by: Mark Grant <m.grant.prg@gmail.com>